+++ RelationalAI.Examples/ListEngines.cs
@@ -2,9 +2,6 @@ using System;
 using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
 using System.Threading.Tasks;
-using RelationalAI.Models.Engine;
-using RelationalAI.Services;
-using RelationalAI.Utils;
 
 namespace RelationalAI.Examples
 {
@@ -13,7 +10,7 @@ namespace RelationalAI.Examples
         public static Command GetCommand()
         {
             var cmd = new Command("ListEngines", "--state <State> --profile <Profile name>"){
-                new Option<EngineState>("--state"){
+                new Option<string>("--state"){
                     IsRequired = false,
                     Description = "To list engines in a paricular state. For example, DELETED"
                 },
@@ -24,11 +21,11 @@ namespace RelationalAI.Examples
                 }
             };
             cmd.Description = "Lists engines.";
-            cmd.Handler = CommandHandler.Create<EngineState?, string>(Run);
+            cmd.Handler = CommandHandler.Create<string, string>(Run);
             return cmd;
         }
 
-        private static async Task Run(EngineState? state = null, string profile = "default")
+        private static async Task Run(string state = null, string profile = "default")
         {
             var config = Config.Read("", profile);
             var context = new Client.Context(config);