+++ RelationalAI.Test/EngineTest.cs
@@ -1,7 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using RelationalAI.Errors;
-using RelationalAI.Models.Engine;
 using Xunit;
 
 namespace RelationalAI.Test
@@ -18,30 +16,23 @@ namespace RelationalAI.Test
 
             var createRsp = await client.CreateEngineWaitAsync(EngineName);
             Assert.Equal(createRsp.Name, EngineName);
-            Assert.Equal(EngineState.Provisioned, createRsp.State);
+            Assert.Equal(EngineStates.Provisioned, createRsp.State);
 
             var engine = await client.GetEngineAsync(EngineName);
             Assert.Equal(engine.Name, EngineName);
-            Assert.Equal(EngineState.Provisioned, engine.State);
+            Assert.Equal(EngineStates.Provisioned, engine.State);
 
             var engines = await client.ListEnginesAsync();
             engine = engines.Find(item => item.Name.Equals(EngineName));
             Assert.NotNull(engine);
 
-            engines = await client.ListEnginesAsync(EngineState.Provisioned);
+            engines = await client.ListEnginesAsync(EngineStates.Provisioned);
             engine = engines.Find(item => item.Name.Equals(EngineName));
             Assert.NotNull(engine);
-
-            await Assert.ThrowsAsync<ArgumentOutOfRangeException>(() =>
-                client.ListEnginesAsync((EngineState)1500));
 
             await Assert.ThrowsAsync<NotFoundException>(async () => await client.DeleteEngineWaitAsync(EngineName));
 
             await Assert.ThrowsAsync<NotFoundException>(async () => await client.GetEngineAsync(EngineName));
-
-            engines = await client.ListEnginesAsync();
-            engine = engines.Find(item => item.Name.Equals(EngineName));
-            Assert.Equal(EngineState.Deleted, engine.State);
         }
 
         public override async Task DisposeAsync()